+++ packages/docusaurus/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/docusaurus
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@0.12.0
+
 ## 0.2.5
 
 ### Patch Changes